pyproject.toml
[project]
name = "python-project-52"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "flake8>=7.2.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "uv>=0.7.3",
    "whitenoise>=6.9.0",
]
